requirements-dev.txt
aiofiles==0.8.0 ; python_version >= "3.8" and python_version < "3.11"
aiohttp==3.8.1 ; python_version >= "3.8" and python_version < "3.11"
aiosignal==1.2.0 ; python_version >= "3.8" and python_version < "3.11"
aiosqlite==0.17.0 ; python_version >= "3.8" and python_version < "3.11"
alembic==1.7.7 ; python_version >= "3.8" and python_version < "3.11"
astroid==2.12.11 ; python_version >= "3.8" and python_version < "3.11"
async-timeout==4.0.2 ; python_version >= "3.8" and python_version < "3.11"
asyncmy==0.2.7 ; python_version >= "3.8" and python_version < "3.11"
atomicwrites==1.4.1 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "win32"
attrs==21.4.0 ; python_version >= "3.8" and python_version < "3.11"
black==22.3.0 ; python_version >= "3.8" and python_version < "3.11"
blinker==1.4 ; python_version >= "3.8" and python_version < "3.11"
cfel-pylint-checkers==1.0.2 ; python_version >= "3.8" and python_version < "3.11"
cffi==1.15.1 ; python_version >= "3.8" and python_version < "3.11"
charset-normalizer==2.1.0 ; python_version >= "3.8" and python_version < "3.11"
click==8.1.3 ; python_version >= "3.8" and python_version < "3.11"
colorama==0.4.5 ; python_version >= "3.8" and python_version < "3.11" and sys_platform == "win32" or python_version >= "3.8" and python_version < "3.11" and platform_system == "Windows"
coverage==6.4.2 ; python_version >= "3.8" and python_version < "3.11"
cryptography==3.4.8 ; python_version >= "3.8" and python_version < "3.11"
dill==0.3.5.1 ; python_version >= "3.8" and python_version < "3.11"
et-xmlfile==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
fawltydeps==0.7.1 ; python_version >= "3.8" and python_version < "3.11"
fire==0.4.0 ; python_version >= "3.8" and python_version < "3.11"
frozenlist==1.3.0 ; python_version >= "3.8" and python_version < "3.11"
greenlet==1.1.2 ; python_version >= "3.8" and python_version < "3.11"
h11==0.13.0 ; python_version >= "3.8" and python_version < "3.11"
h2==4.1.0 ; python_version >= "3.8" and python_version < "3.11"
h5py==3.7.0 ; python_version >= "3.8" and python_version < "3.11"
hpack==4.0.0 ; python_version >= "3.8" and python_version < "3.11"
hypercorn==0.13.2 ; python_version >= "3.8" and python_version < "3.11"
hyperframe==6.0.1 ; python_version >= "3.8" and python_version < "3.11"
hypothesis==6.27.1 ; python_version >= "3.8" and python_version < "3.11"
idna==3.3 ; python_version >= "3.8" and python_version < "3.11"
importlib-metadata==6.1.0 ; python_version >= "3.8" and python_version < "3.11"
importlib-resources==5.8.0 ; python_version >= "3.8" and python_version < "3.9"
iniconfig==1.1.1 ; python_version >= "3.8" and python_version < "3.11"
isort==5.12.0 ; python_version >= "3.8" and python_version < "3.11"
itsdangerous==2.1.2 ; python_version >= "3.8" and python_version < "3.11"
jinja2==3.1.2 ; python_version >= "3.8" and python_version < "3.11"
lark==1.1.2 ; python_version >= "3.8" and python_version < "3.11"
lazy-object-proxy==1.7.1 ; python_version >= "3.8" and python_version < "3.11"
mako==1.2.1 ; python_version >= "3.8" and python_version < "3.11"
markupsafe==2.1.1 ; python_version >= "3.8" and python_version < "3.11"
mccabe==0.7.0 ; python_version >= "3.8" and python_version < "3.11"
multidict==6.0.2 ; python_version >= "3.8" and python_version < "3.11"
mypy-extensions==0.4.3 ; python_version >= "3.8" and python_version < "3.11"
mypy==0.971 ; python_version >= "3.8" and python_version < "3.11"
numpy==1.21.4 ; python_version >= "3.8" and python_version < "3.11"
openpyxl-stubs==0.1.21 ; python_version >= "3.8" and python_version < "3.11"
openpyxl==3.0.9 ; python_version >= "3.8" and python_version < "3.11"
packaging==21.3 ; python_version >= "3.8" and python_version < "3.11"
pathspec==0.10.1 ; python_version >= "3.8" and python_version < "3.11"
pillow==9.1.0 ; python_version >= "3.8" and python_version < "3.11"
pint==0.16.1 ; python_version >= "3.8" and python_version < "3.11"
pip-requirements-parser==32.0.1 ; python_version >= "3.8" and python_version < "3.11"
platformdirs==2.5.2 ; python_version >= "3.8" and python_version < "3.11"
pluggy==1.0.0 ; python_version >= "3.8" and python_version < "3.11"
priority==2.0.0 ; python_version >= "3.8" and python_version < "3.11"
py==1.11.0 ; python_version >= "3.8" and python_version < "3.11"
pycparser==2.21 ; python_version >= "3.8" and python_version < "3.11"
pydantic==1.10.7 ; python_version >= "3.8" and python_version < "3.11"
pyfakefs==4.7.0 ; python_version >= "3.8" and python_version < "3.11"
pylint==2.15.4 ; python_version >= "3.8" and python_version < "3.11"
pyparsing==3.0.9 ; python_version >= "3.8" and python_version < "3.11"
pytest-asyncio==0.17.2 ; python_version >= "3.8" and python_version < "3.11"
pytest-cov==2.12.1 ; python_version >= "3.8" and python_version < "3.11"
pytest-subprocess==1.4.1 ; python_version >= "3.8" and python_version < "3.11"
pytest==6.2.5 ; python_version >= "3.8" and python_version < "3.11"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "3.11"
python-magic==0.4.24 ; python_version >= "3.8" and python_version < "3.11"
pytz==2021.3 ; python_version >= "3.8" and python_version < "3.11"
pyzmq==22.3.0 ; python_version >= "3.8" and python_version < "3.11"
quart-cors==0.5.0 ; python_version >= "3.8" and python_version < "3.11"
quart==0.17.0 ; python_version >= "3.8" and python_version < "3.11"
randomname==0.1.5 ; python_version >= "3.8" and python_version < "3.11"
six==1.16.0 ; python_version >= "3.8" and python_version < "3.11"
sortedcontainers==2.4.0 ; python_version >= "3.8" and python_version < "3.11"
sqlalchemy2-stubs==0.0.2a24 ; python_version >= "3.8" and python_version < "3.11"
sqlalchemy==1.4.31 ; python_version >= "3.8" and python_version < "3.11"
sqlalchemy[asyncio]==1.4.31 ; python_version >= "3.8" and python_version < "3.11"
structlog-overtime==1.0.1 ; python_version >= "3.8" and python_version < "3.11"
structlog==22.1.0 ; python_version >= "3.8" and python_version < "3.11"
termcolor==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
toml==0.10.2 ; python_version >= "3.8" and python_version < "3.11"
tomli==2.0.1 ; python_version >= "3.8" and python_version < "3.11"
tomlkit==0.11.1 ; python_version >= "3.8" and python_version < "3.11"
typed-argument-parser==1.7.2 ; python_version >= "3.8" and python_version < "3.11"
types-python-dateutil==2.8.3 ; python_version >= "3.8" and python_version < "3.11"
types-pytz==2021.3.6 ; python_version >= "3.8" and python_version < "3.11"
types-pyyaml==5.4.12 ; python_version >= "3.8" and python_version < "3.11"
types-requests==2.26.0 ; python_version >= "3.8" and python_version < "3.11"
typing-extensions==4.3.0 ; python_version >= "3.8" and python_version < "3.11"
typing-inspect==0.7.1 ; python_version >= "3.8" and python_version < "3.11"
werkzeug==2.1.2 ; python_version >= "3.8" and python_version < "3.11"
wrapt==1.14.1 ; python_version >= "3.8" and python_version < "3.11"
wsproto==1.1.0 ; python_version >= "3.8" and python_version < "3.11"
yarl==1.7.2 ; python_version >= "3.8" and python_version < "3.11"
zipp==3.8.1 ; python_version >= "3.8" and python_version < "3.11"
